requirements.txt
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.2.1
distro==1.9.0
exceptiongroup==1.3.0
fastapi==0.115.14
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.26
langchain-core==0.3.68
langchain-openai==0.3.27
langchain-text-splitters==0.3.8
langsmith==0.4.5
Mako==1.3.10
MarkupSafe==3.0.2
openai==1.95.1
orjson==3.10.18
packaging==24.2
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
python-decouple==3.8
python-dotenv==1.1.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.41
SQLAlchemy-Utils==0.41.2
starlette==0.46.2
tenacity==9.1.2
tiktoken==0.9.0
tomli==2.2.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
zstandard==0.23.0
langgraph==0.2.46
python-telegram-bot==22.3
